requirements.txt
flask
python-telegram-bot==20.8
requests
beautifulsoup4
gunicorn
python-dotenv